requirements.txt
playwright>=1.40.0
google-cloud-storage>=2.14.0
